Simple Bonebone.js application (Initial Version)
Implement a list page populated with quotes fetched from gist.

- Fetch the remote list of Quotes
- Paging list in groups of 15
- Filter list by themes

KNOWN ISSUE: The first load of the page is having trouble with applying the
filter. In this revision of the app the filters weren't loading until
the paging navigation was used. After using the paging nav the filters
worked fine.